onekey: Fix for V-USB of ATtiny85 and key scan
